corrected some bugs, finished the annoying pnj

+++ GGJ - 2024/Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(float dmg);
+    void Damage(int dmg);
 }
